tests/wpt/web-platform-tests/dom/nodes/ParentNode-append.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>ParentNode.append</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-parentnode-append">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="pre-insertion-validation-hierarchy.js"></script>
<script>
preInsertionValidateHierarchy("append");

function test_append(node, nodeName) {
    test(function() {
        const parent = node.cloneNode();
        parent.append();
        assert_array_equals(parent.childNodes, []);
    }, nodeName + '.append() without any argument, on a parent having no child.');

    test(function() {
        const parent = node.cloneNode();
        parent.append(null);
        assert_equals(parent.childNodes[0].textContent, 'null');
    }, nodeName + '.append() with null as an argument, on a parent having no child.');

    test(function() {
        const parent = node.cloneNode();
        parent.append(undefined);
        assert_equals(parent.childNodes[0].textContent, 'undefined');
    }, nodeName + '.append() with undefined as an argument, on a parent having no child.');

    test(function() {
        const parent = node.cloneNode();
        parent.append('text');
        assert_equals(parent.childNodes[0].textContent, 'text');
    }, nodeName + '.append() with only text as an argument, on a parent having no child.');

    test(function() {
        const parent = node.cloneNode();
        const x = document.createElement('x');
        parent.append(x);
        assert_array_equals(parent.childNodes, [x]);
    }, nodeName + '.append() with only one element as an argument, on a parent having no child.');

    test(function() {
        const parent = node.cloneNode();
        const child = document.createElement('test');
        parent.appendChild(child);
        parent.append(null);
        assert_equals(parent.childNodes[0], child);
        assert_equals(parent.childNodes[1].textContent, 'null');
    }, nodeName + '.append() with null as an argument, on a parent having a child.');

    test(function() {
        const parent = node.cloneNode();
        const x = document.createElement('x');
        const child = document.createElement('test');
        parent.appendChild(child);
        parent.append(x, 'text');
        assert_equals(parent.childNodes[0], child);
        assert_equals(parent.childNodes[1], x);
        assert_equals(parent.childNodes[2].textContent, 'text');
    }, nodeName + '.append() with one element and text as argument, on a parent having a child.');
}

test_append(document.createElement('div'), 'Element');
test_append(document.createDocumentFragment(), 'DocumentFragment');
</script>
</html>
